MedixLine/requirements.txt
asgiref==3.8.1
Django==4.2.16
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
pillow==10.4.0
psycopg2-binary==2.9.9
pycountry==24.6.1
sqlparse==0.5.1
typing_extensions==4.12.2
pycountry==24.6.1
django-cors-headers==4.4.0
